Allow specification of window position in geometry string
The "--geometry" argument is updated so that position information can be
passed along to xinput_calibrator in addition to the traditional window
height/width. The argument format is now `<width>x<height>[+<x>+<y>]`
where "x" and "y" specify where on the desktop the window should be
positioned. For example, to position a 1024x768 window with an offset of
800 pixels on the X axis and 0 pixels on the Y axis, you could use the
argument `1024x768+800+0`. The ability to position the window on the
desktop allows xinput_calibrator to be placed on non-primary displays.

NOTE: The gtkmm backend assumes the geometry string aligns with a
monitor and that the window manager is smart enough to make the window
fullscreen on it.